pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crispr_rl"
version = "0.1.0"
description = "CRISPR guide RNA design engine with reinforcement learning optimization"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "pyyaml",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.110",
    "uvicorn[standard]",
    "pydantic>=2.0",
    "aiosqlite",
    "httpx",
]
test = [
    "pytest",
    "pytest-asyncio",
]
bio = ["biopython"]
all = [
    "fastapi>=0.110",
    "uvicorn[standard]",
    "pydantic>=2.0",
    "aiosqlite",
    "httpx",
    "pytest",
    "pytest-asyncio",
    "biopython",
]
dev = ["black", "ruff", "mypy"]

[tool.setuptools.packages.find]
where = ["."]
include = ["crispr_rl*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
